FIXED ft_memchr ft_strncmp

+++ ft_strncmp.c
@@ -17,7 +17,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	int	i;
 
 	i = 0;
-	while (s1[i] && s2[i] && i < n)
+	while (s1[i] && s2[i] && i < (int) n)
 	{
 		if (s1[i] != s2[i])
 			return (s1[i] - s2[i]);
